requirement.txt
flask
flask-corsrequests
gunicorn